server-side caching implemented

+++ src/services/notes/repositories/note-repositories.js
@@ -1,11 +1,13 @@
 import { nanoid } from 'nanoid';
 import { Pool } from 'pg';
 import collaborationRepositories from '../../collaborations/repositories/collaboration-repositories.js';
+import CacheService from '../../../cache/redis-service.js';
 
 class NoteRepositories {
   constructor() {
     this.pool = new Pool();
     this.collaborationRepositories = collaborationRepositories;
+    this.cacheService = new CacheService();
   }
 
   async createNote({ title, body, tags, owner }) {
@@ -20,21 +22,33 @@ class NoteRepositories {
 
     const result = await this.pool.query(query);
 
+    await this.cacheService.delete(`notes:${owner}`);
+
     return result.rows[0];
   }
 
   async getNotes(owner) {
-    const query = {
-      text: `
-        SELECT notes.* FROM notes 
-        LEFT JOIN collaborations ON collaborations.note_id = notes.id 
-        WHERE notes.owner = $1 OR collaborations.user_id = $1 
-        GROUP BY notes.id
-      `,
-      values: [owner]
+    const cacheKey = `notes:${owner}`;
+
+    try {
+      const notes = await this.cacheService.get(cacheKey);
+      return JSON.parse(notes);
+    } catch {
+      const query = {
+        text: `
+          SELECT notes.* FROM notes 
+          LEFT JOIN collaborations ON collaborations.note_id = notes.id 
+          WHERE notes.owner = $1 OR collaborations.user_id = $1 
+          GROUP BY notes.id
+        `,
+        values: [owner]
+      }
+      const { rows } = await this.pool.query(query);
+
+      await this.cacheService.set(cacheKey, JSON.stringify(rows));
+
+      return rows;
     }
-    const result = await this.pool.query(query);
-    return result.rows;
   }
 
   async getNoteById(id) {
@@ -56,23 +70,34 @@ class NoteRepositories {
     const updatedAt = new Date().toISOString();
 
     const query = {
-      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id',
+      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner',
       values: [title, body, tags, updatedAt, id],
     };
 
-    const result = await this.pool.query(query);
-    return result.rows[0];
+    const { rows } = await this.pool.query(query);
+
+    const owner = rows[0].owner;
+    if (rows[0]) {
+      await this.cacheService.delete(`notes:${owner}`);
+    }
+
+    return rows[0];
   }
 
   async deleteNote(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
-    const result = await this.pool.query(query);
+    const { rows } = await this.pool.query(query);
 
-    return result.rows[0];
+    const owner = rows[0].owner;
+    if (rows[0]) {
+      await this.cacheService.delete(`notes:${owner}`);
+    }
+
+    return rows[0];
   }
 
   async verifyNoteOwner(id, owner) {
